refactor(product): ProductCreateRequestDto에 있던 Product 도메인 의존성 삭제
- 왜: shared에 product 도메인에 대한 의존성이 있기 때문에
- 무엇: ProductCreateRequestDto에 있던 toEntity 메서드 삭제
