+++ src/main/java/com/bugzero/rarego/shared/product/dto/ProductCreateRequestDto.java
@@ -1,14 +1,8 @@
 package com.bugzero.rarego.shared.product.dto;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import com.bugzero.rarego.boundedContext.product.domain.Product;
-import com.bugzero.rarego.boundedContext.product.domain.ProductMember;
-import com.bugzero.rarego.boundedContext.product.domain.dto.ProductImageRequestDto;
 import com.bugzero.rarego.shared.product.type.Category;
-import com.bugzero.rarego.shared.product.type.InspectionStatus;
-import com.bugzero.rarego.shared.product.type.ProductCondition;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
@@ -28,16 +22,4 @@ public record ProductCreateRequestDto(
 	@NotEmpty(message = "이미지는 최소 1장 이상 등록해야 합니다.")
 	List<@Valid ProductImageRequestDto> productImageRequestDto
 ) {
-	public Product toEntity(ProductMember seller) {
-		return Product.builder()
-			.seller(seller)
-			.category(category)
-			.productCondition(ProductCondition.INSPECTION)
-			// 처음 상품이 등록될 때는 검수 대기상태로 지정
-			.inspectionStatus(InspectionStatus.PENDING)
-			.images(new ArrayList<>())
-			.name(name)
-			.description(description)
-			.build();
-	}
 }
