+++ src/main/java/com/bugzero/rarego/shared/product/dto/ProductUpdateDto.java
@@ -2,7 +2,6 @@ package com.bugzero.rarego.shared.product.dto;
 
 import java.util.List;
 
-import com.bugzero.rarego.boundedContext.product.domain.dto.ProductImageUpdateDto;
 import com.bugzero.rarego.shared.product.type.Category;
 
 import jakarta.validation.Valid;
